t8dev.cli.p2a2bin: Fix signedness to handle files loading above $7FFF
We were using a a signed pack that throws an error on values above
$7FFF, meaning that if our object file had a start address at or above
$8000 this would fail. Change it to unsigned so that this works for
those higher-loaded files.

Though we change the length byte also to be written as unsigned, since
that seems more clear, that can't cause an error anyway because the
maximum length for an Apple II `B` file is 32K, so an earlier check
for that would fail before we ever tried to write a length ≥ $8000.