`t8dev pytest`: Collect $T8_PROJDIR when given no file/dir args
This used to try to collect just under src/ and **/{psrc,pylib}/
directories, which isn't really all that desirable as it encodes
a naming convention that not even we really want to keep using.
Pytest does a fine job with its own built-in excludes (especially
by excluding `.*`, which avoids traversing our .build/ directory),
and we generally want to run tests in all Git submodules (Python
modules we don't want to change would be installed from PyPI) so
there's no reason not to search them completely.

Version number bumped to 0.3.0.dev2.