*/psrc/ → */src/: Rename Python source code dirs to be more standard
Since r8format/ and t8dev/ are now (mostly) pure Python packages,
there's no reason to keep non-Python sources separate from the Python
ones. (We do have assembly language source in the packages as well,
but those are actually files distributed with the Python package and
found as "resource" files within it.)

Not all things mentioning "psrc" have been changed; some of these
shouldn't be talking about it in the first place and need to be
fixed in other ways.

This change should be invisible to users, but just in case issues
arise with this, the version numbers are bumped to `.dev3` to make
debugging easier.