r8format: Use importlib_resources for access to test files
We update the version to .dev4 to mark the change in dependencies.

This package now requires the importlib_resources package, which in turn
raises our Python version requirement to 3.7. We use the external
backwards-compatibility package importlib_resources rather than the
built-in importlib.resources because the latter has rather varying APIs
across 3.7 through 3.13, and we don't want to have to deal with those
variations.

This gets rid of binary.pylib() and various other direct path usage.