r8format/programs/ → bastok.programs.msx module
The bastok "functional" tests have been moved from r8format/Functest
and turned into pytest tests in bastok.programs.msx.FUNCTEST.

These tests rely on previous modifications to the bastok.cli programs
to allow calling main() with additional arguments to specify the argv
and the input stream so we don't have to start a separate process to
invoke the program in a way very similar to the way it works from the
command line.

One thing this does not check is the stdin handling, which would be
nice to check because it's sometimes a bit tricky. But for this we
probably want to extract the code that figures out if we're reading
input_override, stdin or a file whose name we've been given and
test that independently.