bastok.cli.detok: Allow specifying different argv1/input to main()
For testing, rather than run the program as a subprocess, we can just
call main() with our list of arguments. We also can specify an
`input_override' stream which will be used instead of reading stdin or
the given filename, making it easy to give it input from a resource
file in the Python package.