t8dev.cli.shared.vprint(): Moved from t8dev.util
This really belongs in cli.shared because thats where the data on
which it depends lives. And bringing it in from there brings in a lot
less than bringing in the cli.util module whence many others were
importing it.

This also modifies t8dev.run.tool() to vprint() the command it's
running because it should do that; we dont' take out the call in
util.runtool because that's going away soon anyway.