t8dev.cli..asl.runasl(): Better error message on assembly failure
runasl() is operating on a different source file from the one supplied
by the user; generally the source file it generates includes the user
code. This was making errors a bit mysterious because you'd get only a
filename, rather than a path to the original source file, and that
filename might be a common one (e.g. "hello") used in several
different subdirectories of exe/ or wherever.

Client of runasl() now pass in a `Path` to the original source file,
runasl() uses `Path.stem` for the name of the generated source file,
but prints an error including the full original `Path` if the assembly
fails, giving the user a much better idea of what's broken.

t8dev.run.tool() has a new argument `errexit=True` that can be
changed to false if the caller wants the return value in order to
print further information about an error, as runasl() does.