bin/z80dasm-clean → intel-disassm-clean: Extend to disx output as well
This improves the number reformatting to get all the numbers on the
line (not just the first one), to handle Intel-style 'nH' numbers
without a leading zero (z80dasm added zeros even when not necessary),
and handle upper-case as well as lower-case trailing 'h'. It's
probably a bit more precise about numbers as well since it now adds
word boundaries to the regular expression match for them.

It converts upper-case mnemonics to lower case, and also upper-case
register names to lower case, mostly. (It misses some relatively rare
ones such as SP, IX and IY, because this doesn't seem worth the effort
to add, at least not right now.)

The re-indentation now handles tabs after the mnemonic; z80dasm used
just spaces.

This has been tested against just one z80dasm output file, but the
ouput is nearly the same. There are a few improvements, e.g. `rst 28h`
→ `rst $28`, and a few lines that have arbitrary indentation changes
on the address comments that we are not too concerned about.

It's also been tested against one disx output file which looks good
enough for the moment.