t8dev bin/cmpasl: Use first-found $bin/asl, not last one
In testing, it was finding the the $projdir/.build/tool/bin/asl and
thus prefixing $projdir/.build/tool/bin to the path, but then because
a `break` was missing it was going and and further prefixing other
versions that it found. This broke code that depended on getting the
latest bin/asl version in $projdir, but instead was using older
versions found in e.g., /opt/asl/bin.