t8dev//cmpasl: Add -f 0xNN option for fill byte
Depending on the original ROM image, large areas for which we prefer
not to generate code may be filled with $00 or $FF. Apparently it's
been more often the former so far, but EPROM images often tend to
prefer the latter since it allows reprogramming of just those bytes.

This can be worked around by directly filling areas (`db nn dup $FF`)
in the reverse-engienered source code, but that's a pain.