t8dev.toolset.dos33fs → dos33fsprogs renaming
The repo itself was renamed from dos33fs to dos33fsprogs quite some
time ago, but we never changed our local name to match. This is
getting confusing, so no we change the local name.

This will break anybody using `t8dev toolset dos33fs`, but it's a
simple enough fix for them that it doesn't seem worth installing an
alias.