r8format binary.tool.intelhex: Generate Intel HEX records from a MemImage
This works the same way as ASL's `p2hex` program: it generates
hex records in the order they appear in the .p file (though this
does not warn about overlaps).

The aslhex.xxd file is an xxd dump of a .p file generated with
overlapping records with an earlier address in a later record.
`xxd -r` will produce the binary .p file that can then be used to
confirm `p2hex` output.