t8dev.toolset.asl: Fix check_installed() for latest versions
We were previously passing -this-is-not-an-option and expecting an
error return with 'Invalid option' in the output. We did this because
the -version parameter didn't exist at the time we added the original
check.

However, ASL recently changed the error message above to be 'invalid
parameter', which broke the above check. Since we must change it
anyway, we might as well switch to using -version now since it's been
around for long enough that few people are probably using such an old
version of ASL that it doesn't have this.

It's safe to include the version '1.42' in the output check becuase
this appears unlikely ever to change: it's been 1.42 for many years
now with only the 'Bld' number (currently 292) incrementing.